Stand-alone version.
So here it is, the stand-alone version (that is also installed in the
Github pages branch so people can see it live on the 'tubes.)

At this point the Flask-hosted version is slightly less interesting (not
in an absolute sense, just for the nonce) because there are so many
gnarly possibilities in the self-hosted JS version of Xerblin.

We can create any old command word to run, say, D3 animations, or to
create new SVG elements, or do anything with jQuery...  The mind fairly
boggles!